apply Threshold modifications

+++ run_classification.py
@@ -130,7 +130,7 @@ def evaluate(lfd_params, model, mode="evaluation", verbose=False, input_dtype="v
 
     for i, data_packet in enumerate(data_loader):
         obs, label, filename = data_packet
-
+        obs = obs.float()
         print("obs:", obs.shape)
 
         # compute output
